backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.5.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
packaging==25.0
passlib==1.7.4
pillow==12.1.1
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
psutil==7.2.2
psycopg2-binary==2.9.11
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.24
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
langchain==1.2.13
langchain-core==1.2.23
langsmith==0.7.23
langchain-groq==1.1.2
groq==0.37.1
distro==1.9.0
sniffio==1.3.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
jsonpatch==1.33
jsonpointer==3.1.1
tenacity==9.1.4
uuid-utils==0.14.1
xxhash==3.6.0
orjson==3.11.7
ormsgpack==1.12.2
zstandard==0.25.0
reportlab==4.4.4
imageio-ffmpeg
huggingface_hub